apps/tymble-api/src/drizzle/seed/instruments/requirements.txt
yfinance>=0.2.40
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
